refactor: split GeometryUtils into Tier 1 (engine-api) and Tier 2 (zones)
Eliminates the architectural problem where phosphor-tiles would need to
link phosphor-zones just for pure geometry functions, creating unwanted
cross-engine coupling.

Tier 1 — PhosphorEngineApi::GeometryUtils (no Zone/Layout deps):
  enforceWindowMinSizes, removeZoneOverlaps, snapToRect, rectToJson,
  serializeZoneAssignments, availableAreaToOverlayCoordinates

Tier 2 — PhosphorZones::GeometryUtils (Zone/Layout-aware):
  getZoneGeometryWithGaps, getZoneGeometryForScreen(F),
  getEffectiveZonePadding, getEffectiveOuterGaps,
  effectiveScreenGeometry, extractZoneGeometry, setZoneGeometry

PhosphorZones::GeometryUtils re-exports Tier 1 via using-declarations,
so existing callers using the PhosphorZones namespace see both tiers.

Dependency graph is now clean:
  engine-api ← zones (Zone-aware geometry)
  engine-api ← tiles (AutotileEngine uses enforceWindowMinSizes)
  No zones ↔ tiles dependency.
