+++ libs/phosphor-zones/include/PhosphorZones/GeometryUtils.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <PhosphorEngineApi/EngineTypes.h>
+#include <PhosphorEngineApi/GeometryUtils.h>
 #include <PhosphorEngineApi/IGeometrySettings.h>
 #include <phosphorzones_export.h>
 
@@ -11,11 +11,8 @@
 
 #include <QRect>
 #include <QRectF>
-#include <QSize>
 #include <QString>
-#include <QVector>
-
-#include <functional>
+#include <QVariantMap>
 
 class QScreen;
 
@@ -31,8 +28,12 @@ class ScreenManager;
 namespace PhosphorZones {
 namespace GeometryUtils {
 
-PHOSPHORZONES_EXPORT QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, QScreen* screen);
-PHOSPHORZONES_EXPORT QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, const QRect& overlayGeometry);
+using PhosphorEngineApi::GeometryUtils::availableAreaToOverlayCoordinates;
+using PhosphorEngineApi::GeometryUtils::enforceWindowMinSizes;
+using PhosphorEngineApi::GeometryUtils::rectToJson;
+using PhosphorEngineApi::GeometryUtils::removeZoneOverlaps;
+using PhosphorEngineApi::GeometryUtils::serializeZoneAssignments;
+using PhosphorEngineApi::GeometryUtils::snapToRect;
 
 PHOSPHORZONES_EXPORT QRectF getZoneGeometryWithGaps(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone,
                                                     QScreen* screen, int innerGap,
@@ -58,8 +59,6 @@ PHOSPHORZONES_EXPORT int getEffectiveZonePadding(PhosphorZones::Layout* layout,
                                                  PhosphorEngineApi::IGeometrySettings* settings,
                                                  const QString& screenId = {});
 
-PHOSPHORZONES_EXPORT QRect snapToRect(const QRectF& rf);
-
 PHOSPHORZONES_EXPORT ::PhosphorLayout::EdgeGaps getEffectiveOuterGaps(PhosphorZones::Layout* layout,
                                                                       PhosphorEngineApi::IGeometrySettings* settings,
                                                                       const QString& screenId = {});
@@ -72,16 +71,6 @@ PHOSPHORZONES_EXPORT QRectF effectiveScreenGeometry(Phosphor::Screens::ScreenMan
 
 PHOSPHORZONES_EXPORT QRectF extractZoneGeometry(const QVariantMap& zone);
 PHOSPHORZONES_EXPORT void setZoneGeometry(QVariantMap& zone, const QRectF& rect);
-
-PHOSPHORZONES_EXPORT void enforceWindowMinSizes(QVector<QRect>& zones, const QVector<QSize>& minSizes, int gapThreshold,
-                                                int innerGap = 0);
-
-PHOSPHORZONES_EXPORT void removeZoneOverlaps(QVector<QRect>& zones, const QVector<QSize>& minSizes = {},
-                                             int innerGap = 0);
-
-PHOSPHORZONES_EXPORT QString rectToJson(const QRect& rect);
-
-PHOSPHORZONES_EXPORT QString serializeZoneAssignments(const QVector<PhosphorEngineApi::ZoneAssignmentEntry>& entries);
 
 } // namespace GeometryUtils
 } // namespace PhosphorZones
