+++ src/core/geometryutils.cpp
@@ -4,26 +4,18 @@
 #include "geometryutils.h"
 #include <PhosphorZones/Zone.h>
 #include <geometry_helpers.h>
-#include "logging.h"
 #include <PhosphorZones/Layout.h>
 #include "layoutworker/layoutcomputeservice.h"
 #include "interfaces.h"
 #include "constants.h"
 #include <PhosphorScreens/Manager.h>
-#include "utils.h"
-#include <algorithm>
-#include <QJsonArray>
-#include <QJsonDocument>
-#include <QJsonObject>
 #include <QScreen>
 #include <QVariantMap>
 #include <PhosphorScreens/ScreenIdentity.h>
+#include <PhosphorZones/ZoneDefaults.h>
 
 namespace PlasmaZones {
 
-/// Get per-screen snapping settings with virtual screen fallback.
-/// Tries the exact screenId first; if empty and the screenId is a virtual
-/// screen, falls back to the physical parent's overrides.
 static QVariantMap getPerScreenSnappingWithFallback(ISettings* settings, const QString& screenId)
 {
     QVariantMap result = settings->getPerScreenSnappingSettings(screenId);
@@ -51,189 +43,6 @@ ScreenGeometries resolveScreenGeometries(Phosphor::Screens::ScreenManager* mgr, 
 }
 } // anonymous namespace
 
-static QRectF calculateZoneGeometry(PhosphorZones::Zone* zone, QScreen* screen)
-{
-    if (!zone || !screen) {
-        return QRectF();
-    }
-
-    const QRectF screenGeom = screen->geometry();
-    return zone->calculateAbsoluteGeometry(screenGeom);
-}
-
-QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, QScreen* screen)
-{
-    if (!screen) {
-        return geometry;
-    }
-
-    // The overlay window covers the full screen (geometry()).
-    // The geometry parameter is already in absolute screen coordinates
-    // (from available or full-screen geometry), so we just need to convert
-    // to overlay-local coordinates by subtracting the full screen origin.
-    const QRectF screenGeom = screen->geometry();
-    return QRectF(geometry.x() - screenGeom.x(), geometry.y() - screenGeom.y(), geometry.width(), geometry.height());
-}
-
-static QRectF calculateZoneGeometryInAvailableArea(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone,
-                                                   QScreen* screen)
-{
-    if (!zone || !screen) {
-        return QRectF();
-    }
-
-    // Use ScreenManager's panel-aware available geometry when present; fall
-    // back to Qt's availableGeometry() (no-panel environments, unit tests).
-    const QRectF availableGeom = mgr ? mgr->actualAvailableGeometry(screen) : screen->availableGeometry();
-    return zone->calculateAbsoluteGeometry(availableGeom);
-}
-
-/**
- * @brief Detect whether each edge of a zone lies at a screen boundary
- * @param zone The zone to check
- * @param screenGeom The reference screen geometry (for fixed mode pixel checks)
- * @return Array of 4 bools: [left, top, right, bottom] — true if at boundary
- */
-struct EdgeBoundaries
-{
-    bool left = false;
-    bool top = false;
-    bool right = false;
-    bool bottom = false;
-};
-
-static EdgeBoundaries detectEdgeBoundaries(PhosphorZones::Zone* zone, const QRectF& screenGeom)
-{
-    EdgeBoundaries edges;
-
-    if (zone->isFixedGeometry()) {
-        // Fixed mode: pixel proximity check (within 5px of screen boundary)
-        constexpr qreal pixelTolerance = 5.0;
-        QRectF fixedGeo = zone->fixedGeometry();
-        edges.left = (fixedGeo.left() < pixelTolerance);
-        edges.top = (fixedGeo.top() < pixelTolerance);
-        edges.right = (fixedGeo.right() > (screenGeom.width() - pixelTolerance));
-        edges.bottom = (fixedGeo.bottom() > (screenGeom.height() - pixelTolerance));
-    } else {
-        // Relative mode: existing check (near 0 or 1, tolerance 0.01)
-        constexpr qreal edgeTolerance = 0.01;
-        QRectF relGeom = zone->relativeGeometry();
-        edges.left = (relGeom.left() < edgeTolerance);
-        edges.top = (relGeom.top() < edgeTolerance);
-        edges.right = (relGeom.right() > (1.0 - edgeTolerance));
-        edges.bottom = (relGeom.bottom() > (1.0 - edgeTolerance));
-    }
-
-    return edges;
-}
-
-/**
- * @brief Apply gap adjustments to a zone's absolute geometry
- * @param zoneGeom Absolute zone geometry (already calculated against reference area)
- * @param zone PhosphorZones::Zone for edge boundary detection
- * @param referenceGeom Reference geometry for fixed-mode edge detection
- * @param innerGap Gap between adjacent zones (zonePadding)
- * @param outerGaps Per-side edge gaps
- * @return Geometry with gaps applied
- *
- * Shared helper used by both QScreen* and QRect overloads of getZoneGeometryWithGaps.
- */
-static QRectF applyGapsToZoneGeometry(const QRectF& zoneGeom, PhosphorZones::Zone* zone, const QRectF& referenceGeom,
-                                      int innerGap, const ::PhosphorLayout::EdgeGaps& outerGaps,
-                                      const Phosphor::Screens::VirtualScreenDef::PhysicalEdges& physEdges = {
-                                          true, true, true, true})
-{
-    // Detect which edges are at screen boundaries
-    EdgeBoundaries edges = detectEdgeBoundaries(zone, referenceGeom);
-
-    // Only apply outer gaps on edges that are at the PHYSICAL screen boundary.
-    // Internal edges (shared with another virtual screen) get inner gap to avoid
-    // double gaps.
-    qreal leftAdj = (edges.left && physEdges.left) ? outerGaps.left : (innerGap / 2.0);
-    qreal topAdj = (edges.top && physEdges.top) ? outerGaps.top : (innerGap / 2.0);
-    qreal rightAdj = (edges.right && physEdges.right) ? outerGaps.right : (innerGap / 2.0);
-    qreal bottomAdj = (edges.bottom && physEdges.bottom) ? outerGaps.bottom : (innerGap / 2.0);
-
-    // Apply the adjustments (positive inset from edges)
-    QRectF geoF = zoneGeom.adjusted(leftAdj, topAdj, -rightAdj, -bottomAdj);
-
-    // Ensure gaps don't collapse the zone to negative dimensions
-    if (geoF.width() < 1.0) {
-        geoF.setWidth(1.0);
-    }
-    if (geoF.height() < 1.0) {
-        geoF.setHeight(1.0);
-    }
-
-    return geoF;
-}
-
-QRectF getZoneGeometryWithGaps(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone, QScreen* screen,
-                               int innerGap, const ::PhosphorLayout::EdgeGaps& outerGaps, bool useAvailableGeometry,
-                               const QString& screenId)
-{
-    if (!zone || !screen) {
-        return QRectF();
-    }
-
-    // Use available geometry (excludes panels/taskbars) or full screen geometry
-    QRectF geom;
-    if (useAvailableGeometry) {
-        geom = calculateZoneGeometryInAvailableArea(mgr, zone, screen);
-    } else {
-        geom = calculateZoneGeometry(zone, screen);
-    }
-
-    // Detect which edges are at screen boundaries
-    QRectF screenGeom = useAvailableGeometry
-        ? (mgr ? mgr->actualAvailableGeometry(screen) : screen->availableGeometry())
-        : screen->geometry();
-
-    // Look up physical edges via Phosphor::Screens::ScreenManager so virtual screen internal edges
-    // get inner gap instead of outer gap, matching the QRect overload behavior.
-    // When the caller provides a virtual screen ID, use it; otherwise fall back
-    // to the physical screen identifier (which yields all-true physical edges).
-    Phosphor::Screens::VirtualScreenDef::PhysicalEdges physEdges{true, true, true, true};
-    if (mgr) {
-        QString resolvedId = screenId.isEmpty() ? Phosphor::Screens::ScreenIdentity::identifierFor(screen) : screenId;
-        physEdges = mgr->physicalEdgesFor(resolvedId);
-    }
-
-    return applyGapsToZoneGeometry(geom, zone, screenGeom, innerGap, outerGaps, physEdges);
-}
-
-QRectF availableAreaToOverlayCoordinates(const QRectF& geometry, const QRect& overlayGeometry)
-{
-    return QRectF(geometry.x() - overlayGeometry.x(), geometry.y() - overlayGeometry.y(), geometry.width(),
-                  geometry.height());
-}
-
-QRectF getZoneGeometryWithGaps(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone,
-                               const QRect& screenGeometry, const QRect& availableGeometry, int innerGap,
-                               const ::PhosphorLayout::EdgeGaps& outerGaps, bool useAvailableGeometry,
-                               const QString& screenId)
-{
-    if (!zone) {
-        return QRectF();
-    }
-
-    // Calculate absolute zone geometry against the chosen reference area
-    QRectF referenceGeom = useAvailableGeometry ? QRectF(availableGeometry) : QRectF(screenGeometry);
-    QRectF geom = zone->calculateAbsoluteGeometry(referenceGeom);
-
-    // Determine which edges are at the physical screen boundary.
-    // For virtual screens, internal edges (shared with another virtual screen) get inner
-    // gap instead of outer gap to avoid double gaps at virtual screen boundaries.
-    // For physical screens, physicalEdgesFor() returns all-true (all edges are outer).
-    // When @p mgr is null (unit tests), applies default all-true physEdges.
-    if (!screenId.isEmpty() && mgr) {
-        auto pe = mgr->physicalEdgesFor(screenId);
-        return applyGapsToZoneGeometry(geom, zone, referenceGeom, innerGap, outerGaps, pe);
-    }
-
-    return applyGapsToZoneGeometry(geom, zone, referenceGeom, innerGap, outerGaps);
-}
-
 QRectF getZoneGeometryForScreenF(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Zone* zone, QScreen* screen,
                                  const QString& screenId, PhosphorZones::Layout* layout, ISettings* settings)
 {
@@ -251,14 +60,6 @@ QRectF getZoneGeometryForScreenF(Phosphor::Screens::ScreenManager* mgr, Phosphor
         QRect availGeom = vsAvailGeom.isValid() ? vsAvailGeom : vsGeom;
         return getZoneGeometryWithGaps(mgr, zone, vsGeom, availGeom, zonePadding, outerGaps, useAvail, screenId);
     }
-    // For virtual screen IDs, do NOT fall back to the parent QScreen* — that
-    // re-projects relative zone coordinates (defined for the VS sub-region)
-    // onto the full physical screen, producing a geometry that lands in a
-    // DIFFERENT virtual screen. processBatchEntries then re-derives the
-    // window's screen from the geometry center, silently overwriting the
-    // stored VS scope and corrupting WindowScreenAssignments. Better to
-    // return an invalid rect so the caller skips the entry; the next resnap
-    // (after VS state is restored) will succeed.
     if (PhosphorIdentity::VirtualScreenId::isVirtual(screenId)) {
         return QRectF();
     }
@@ -277,7 +78,6 @@ QRect getZoneGeometryForScreen(Phosphor::Screens::ScreenManager* mgr, PhosphorZo
 
 int getEffectiveZonePadding(PhosphorZones::Layout* layout, ISettings* settings, const QString& screenId)
 {
-    // Per-screen snapping override (highest priority), with virtual screen fallback
     if (!screenId.isEmpty() && settings) {
         QVariantMap perScreen = getPerScreenSnappingWithFallback(settings, screenId);
         auto it = perScreen.constFind(QLatin1String(PerScreenSnappingKey::ZonePadding));
@@ -285,27 +85,18 @@ int getEffectiveZonePadding(PhosphorZones::Layout* layout, ISettings* settings, 
             return it->toInt();
         }
     }
-    // Check for layout-specific override
     if (layout && layout->hasZonePaddingOverride()) {
         return layout->zonePadding();
     }
-    // Fall back to global settings
     if (settings) {
         return settings->zonePadding();
     }
-    // Last resort: use default constant
     return Defaults::ZonePadding;
-}
-
-QRect snapToRect(const QRectF& rf)
-{
-    return GeometryHelpers::snapToRect(rf);
 }
 
 ::PhosphorLayout::EdgeGaps getEffectiveOuterGaps(PhosphorZones::Layout* layout, ISettings* settings,
                                                  const QString& screenId)
 {
-    // Per-screen snapping override (highest priority), with virtual screen fallback
     if (!screenId.isEmpty() && settings) {
         QVariantMap perScreen = getPerScreenSnappingWithFallback(settings, screenId);
         if (!perScreen.isEmpty()) {
@@ -316,10 +107,8 @@ QRect snapToRect(const QRectF& rf)
                 auto bottomIt = perScreen.constFind(QLatin1String(PerScreenSnappingKey::OuterGapBottom));
                 auto leftIt = perScreen.constFind(QLatin1String(PerScreenSnappingKey::OuterGapLeft));
                 auto rightIt = perScreen.constFind(QLatin1String(PerScreenSnappingKey::OuterGapRight));
-                // If any per-side key is present, use per-screen per-side gaps
                 if (topIt != perScreen.constEnd() || bottomIt != perScreen.constEnd() || leftIt != perScreen.constEnd()
                     || rightIt != perScreen.constEnd()) {
-                    // Fall back to per-screen uniform OuterGap, then global for missing sides
                     auto uniformIt = perScreen.constFind(QLatin1String(PerScreenSnappingKey::OuterGap));
                     int fallback = (uniformIt != perScreen.constEnd()) ? uniformIt->toInt() : settings->outerGap();
                     return {(topIt != perScreen.constEnd()) ? topIt->toInt() : fallback,
@@ -328,8 +117,6 @@ QRect snapToRect(const QRectF& rf)
                             (rightIt != perScreen.constEnd()) ? rightIt->toInt() : fallback};
                 }
             }
-            // UsePerSideOuterGap not set or false — per-side keys ignored if present
-            // Per-screen uniform outer gap
             auto uniformIt = perScreen.constFind(QLatin1String(PerScreenSnappingKey::OuterGap));
             if (uniformIt != perScreen.constEnd()) {
                 return ::PhosphorLayout::EdgeGaps::uniform(uniformIt->toInt());
@@ -337,10 +124,8 @@ QRect snapToRect(const QRectF& rf)
         }
     }
 
-    // Check for layout-specific per-side override first
     if (layout && layout->usePerSideOuterGap() && layout->hasPerSideOuterGapOverride()) {
         ::PhosphorLayout::EdgeGaps gaps = layout->rawOuterGaps();
-        // Fill in -1 sentinel values from global per-side or uniform fallback
         if (settings && settings->usePerSideOuterGap()) {
             if (gaps.top < 0)
                 gaps.top = settings->outerGapTop();
@@ -364,12 +149,10 @@ QRect snapToRect(const QRectF& rf)
         return gaps;
     }
 
-    // Check for layout-specific uniform override
     if (layout && layout->hasOuterGapOverride()) {
         return ::PhosphorLayout::EdgeGaps::uniform(layout->outerGap());
     }
 
-    // Fall back to global settings
     if (settings) {
         if (settings->usePerSideOuterGap()) {
             return {settings->outerGapTop(), settings->outerGapBottom(), settings->outerGapLeft(),
@@ -381,51 +164,6 @@ QRect snapToRect(const QRectF& rf)
     return ::PhosphorLayout::EdgeGaps::uniform(Defaults::OuterGap);
 }
 
-QRectF effectiveScreenGeometry(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Layout* layout, QScreen* screen)
-{
-    if (!screen) {
-        return QRectF();
-    }
-    if (layout && layout->useFullScreenGeometry()) {
-        return screen->geometry();
-    }
-    return mgr ? mgr->actualAvailableGeometry(screen) : screen->availableGeometry();
-}
-
-QRectF effectiveScreenGeometry(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Layout* layout,
-                               const QString& screenId)
-{
-    auto [geom, availGeom] = resolveScreenGeometries(mgr, screenId);
-    if (geom.isValid()) {
-        if (layout && layout->useFullScreenGeometry()) {
-            return QRectF(geom);
-        }
-        return availGeom.isValid() ? QRectF(availGeom) : QRectF(geom);
-    }
-    // Fallback to physical screen
-    QScreen* screen = Phosphor::Screens::ScreenIdentity::findByIdOrName(screenId);
-    return effectiveScreenGeometry(mgr, layout, screen);
-}
-
-QRectF extractZoneGeometry(const QVariantMap& zone)
-{
-    return QRectF(zone.value(::PhosphorZones::ZoneJsonKeys::X).toDouble(),
-                  zone.value(::PhosphorZones::ZoneJsonKeys::Y).toDouble(),
-                  zone.value(::PhosphorZones::ZoneJsonKeys::Width).toDouble(),
-                  zone.value(::PhosphorZones::ZoneJsonKeys::Height).toDouble());
-}
-
-void setZoneGeometry(QVariantMap& zone, const QRectF& rect)
-{
-    zone[::PhosphorZones::ZoneJsonKeys::X] = rect.x();
-    zone[::PhosphorZones::ZoneJsonKeys::Y] = rect.y();
-    zone[::PhosphorZones::ZoneJsonKeys::Width] = rect.width();
-    zone[::PhosphorZones::ZoneJsonKeys::Height] = rect.height();
-}
-
-/// Iterates empty zones and builds EmptyZoneEntry values for the typed D-Bus result.
-/// EmptyZoneEntry values. Accepts either explicit screen geometry (virtual screen path)
-/// or a physical QScreen* (physical screen path).
 static EmptyZoneList buildEmptyZoneListImpl(Phosphor::Screens::ScreenManager* mgr, PhosphorZones::Layout* layout,
                                             const std::optional<QRect>& screenGeometry, const QRect& availableGeometry,
                                             const QString& screenId, int zonePadding,
@@ -527,42 +265,6 @@ EmptyZoneList buildEmptyZoneList(Phosphor::Screens::ScreenManager* mgr, Phosphor
     }
 
     return physScreen ? buildEmptyZoneList(mgr, layout, physScreen, settings, isZoneEmpty) : EmptyZoneList{};
-}
-
-QString rectToJson(const QRect& rect)
-{
-    QJsonObject obj;
-    obj[::PhosphorZones::ZoneJsonKeys::X] = rect.x();
-    obj[::PhosphorZones::ZoneJsonKeys::Y] = rect.y();
-    obj[::PhosphorZones::ZoneJsonKeys::Width] = rect.width();
-    obj[::PhosphorZones::ZoneJsonKeys::Height] = rect.height();
-    return QString::fromUtf8(QJsonDocument(obj).toJson(QJsonDocument::Compact));
-}
-
-QString serializeZoneAssignments(const QVector<ZoneAssignmentEntry>& entries)
-{
-    if (entries.isEmpty()) {
-        return QStringLiteral("[]");
-    }
-    QJsonArray array;
-    for (const ZoneAssignmentEntry& entry : entries) {
-        QJsonObject obj;
-        obj[JsonKeys::WindowId] = entry.windowId;
-        obj[JsonKeys::SourceZoneId] = entry.sourceZoneId;
-        obj[JsonKeys::TargetZoneId] = entry.targetZoneId;
-        if (!entry.targetZoneIds.isEmpty()) {
-            QJsonArray zoneIdsArr;
-            for (const QString& zid : entry.targetZoneIds)
-                zoneIdsArr.append(zid);
-            obj[JsonKeys::TargetZoneIds] = zoneIdsArr;
-        }
-        obj[::PhosphorZones::ZoneJsonKeys::X] = entry.targetGeometry.x();
-        obj[::PhosphorZones::ZoneJsonKeys::Y] = entry.targetGeometry.y();
-        obj[::PhosphorZones::ZoneJsonKeys::Width] = entry.targetGeometry.width();
-        obj[::PhosphorZones::ZoneJsonKeys::Height] = entry.targetGeometry.height();
-        array.append(obj);
-    }
-    return QString::fromUtf8(QJsonDocument(array).toJson(QJsonDocument::Compact));
 }
 
 } // namespace GeometryUtils
